--- src/main/java/at/grueneis/spengergasse/registry5/Entity.java
@@ -1,17 +1,0 @@
-package at.grueneis.spengergasse.registry5;
-
-/**
- * Internal class used for saving objects and information
- */
-
-public class Entity {
-	private Object originalObject;
-
-	public Entity(Object objectToSave) {
-		this.originalObject = objectToSave;
-	}
-
-	public Object getObject() {
-		return this.originalObject;
-	}
-}
