--- src/main/java/at/grueneis/spengergasse/registry5/Registry.java
@@ -1,11 +1,0 @@
-package at.grueneis.spengergasse.registry5;
-
-/**
- * Created by Daniel on 22/01/14.
- */
-public class Registry {
-
-	public static void add(Object objectToAdd) {
-
-	}
-}
