+++ src/test/java/at/grueneis/spengergasse/registry/EntityTest.java
@@ -1,4 +1,4 @@
-package at.grueneis.spengergasse.registry5;
+package at.grueneis.spengergasse.registry;
 
 import org.junit.Test;
